+++ src/main/java/roomescape/exception/reservation/ThemeNotFoundException.java
@@ -1,0 +1,10 @@
+package roomescape.exception.reservation;
+
+import roomescape.exception.NotFoundException;
+
+public class ThemeNotFoundException extends NotFoundException {
+
+    public ThemeNotFoundException() {
+        super("테마가 존재하지 않습니다.");
+    }
+}
