--- src/main/java/roomescape/repository/exception/ThemeNotFoundException.java
@@ -1,10 +1,0 @@
-package roomescape.repository.exception;
-
-import roomescape.exception.NotFoundException;
-
-public class ThemeNotFoundException extends NotFoundException {
-
-    public ThemeNotFoundException(String message) {
-        super(message);
-    }
-}
